add-director branch

+++ src/main/java/ru/java/practicum/filmorate/storage/LikesStorage.java
@@ -14,6 +14,6 @@ public interface LikesStorage {
 
     List<Film> getPopularFilms(int count);
 
-    int getLikesCountForFilm(Long filmId);
+    long getLikesCountForFilm(Long filmId);
 
 }
